remove requirements.txt. move agreement to modules.
