--- packages/examples/cvat/recording-oracle/requirements.txt
@@ -1,3 +1,0 @@
-numpy
-
-pytest
